renamed applicationState.dart and authenticationHelpers.dart to match naming convention
