+++ lib/main.dart
@@ -14,7 +14,7 @@ import 'package:firebase_core/firebase_core.dart';
 // generated by flutterfire cli for soundspace
 import 'firebase_options.dart';
 //auth dependency
-import 'pages/account/applicationState.dart';
+import 'pages/account/application_state.dart';
 
 //flutter test
 import 'package:flutter_test/flutter_test.dart';
