+++ lib/pages/account/authentication_helpers.dart
