+++ lib/pages/upload/upload.dart
@@ -5,7 +5,7 @@ import 'music_handler.dart';
 import 'image_handler.dart';
 import 'video_handler.dart';
 //to properly use ApplicationState context
-import 'package:soundspace/pages/account/applicationState.dart';
+import 'package:soundspace/pages/account/application_state.dart';
 import 'package:provider/provider.dart';
 import 'package:soundspace/pages/account/authentication.dart';
 //to redirect to login page (currently "Account")
